+++ src/Facades/ZenzivaFacade.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace TuxDaemon\ZenzivaNotificationChannel\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+class ZenzivaFacade extends Facade
+{
+    protected static function getFacadeAccessor()
+    {
+        return 'zenzivasms';
+    }
+}