+++ CHANGELOG.md
@@ -7,3 +7,11 @@ All notable changes to `daemon144key/zenziva-notification-channel` will be docum
 ### Changes
 
 - Initial implementation of the Zenziva Http client for laravel notification channel
+
+
+## 0.9.7 - 2019-09-14
+
+### Changes
+
+- Add supporting artisan command and facade
+- Add wrapper function to check SMS Credit/Balance